python_helper/requirements.txt
httpx==0.27.0
pydantic==2.8.2
numpy==1.26.4
faiss-cpu==1.8.0
llama-cpp-python>=0.3.16
tiktoken==0.7.0
python-dotenv==1.0.1
